+++ frontend/src/assets/images/gem-found-purple.png
