+++ frontend/src/assets/videos/gem-found-purple.mp4
